Send FicTrac update frame signal to DAQ
Added code for sending next frame signal to the DAQ on a digital
channel. Need to go test it in the lab. Also maid some quality
improvements on the signaling between the different processes.
Aquisition is now the first task to start. Main process waits till it
gets the ready signal. Then FicTrac is started. Main process waits till
it gets the ready signal from FicTrac. Then playback is started. Need to
test in the lab setup now.