video: implement playlist item looping, durations, random modes
now playlist items atomatically advance to their next one when
they define a duration (subclasses may also overrided @duration
and provide it themselves)

playlist advances to the next item in the list according to
the configured randomisation modes and repeat configuration

also implement the video_action verb to play/pause the playlist

(this required changing the process model of the video sever
to be single process because generators are not serializable
and they are heavily used in the randomizer. in any case not
a bad decision as it brings it in line with the audio server.
note: the main launcher is currently broken because of this
change and will need to be fixed)