stimuli: fix amplitude / intensity confusion
we are still carrying around a bit of backwards compatibility
cruft, but tidy this in a consistent way.

intensity is applied to *all* phases of the signal and
can be thought of as a signal-independent scale factor

amplitude is signal/meaning dependent, such as a pulse
signal having two configurable amplitudes.

if the API is consistent (to verify), intensity can be
passed to all stimuli