sounds_server: always use a fixed buffer size
this was the actual behaviour but the docstring suggested
that other possiblities were a good idea - which they are not.

portaudio, if passed a latency target, reserves the right to
have a different buffer size *per callback*, which is
incompatible with chunker being fixed size. so disable
passing anything other than a fixed buffer size