fix: open sempaphore on the python side had a bug.
The code that opens the semaphore on the python side keeps retrying
to open it for a fix number of times (originall 10,000). I thought
this was enough but things take a bit longer to startup it seems on
the test machine. This caused the semaphore not to be opened. This
should have caused an error to be generated but that edge case wasn't
coded\tested properly. I think this fix should handle things better.
Still would be good to add a test, I will have to figure out how to
do that best.